refactor: 스타일 파일명을 더 적절한 이름으로 수정
